backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
